+++ CHANGELOG.rst
@@ -24,7 +24,8 @@ Changed:
 * Changed QC thresholds for WGS normal and WES https://github.com/Clinical-Genomics/BALSAMIC/pull/1477
 * Change VarDict memory usage to fix crashes in production https://github.com/Clinical-Genomics/BALSAMIC/pull/1537
 * Updated cluster resources for tnscope WGS TN https://github.com/Clinical-Genomics/BALSAMIC/pull/1535
-* changed --pcr_indel_model to conservative for TNscope in all TGA workflows https://github.com/Clinical-Genomics/BALSAMIC/pull/1526
+* Disable SV calling in TNscope https://github.com/Clinical-Genomics/BALSAMIC/pull/1541
+
 
 Removed:
 ^^^^^^^^